requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
setuptools>=70.0.0
whitenoise==6.6.0
drf-yasg==1.21.7

